#1339 - Modularization of Action Bar
- Factored document export action bar item
