+++ webanno-ui-curation/src/main/java/de/tudarmstadt/ukp/clarin/webanno/ui/curation/page/CurationPage.html
@@ -95,7 +95,6 @@
   <wicket:extend>
     <div wicket:id=openDocumentsModal></div>
     <div wicket:id=remergeDocumentDialog></div>
-    <div wicket:id="exportDialog"></div>
     <div wicket:id="finishDocumentDialog"></div>
 
     <div class="actionbar">
@@ -109,9 +108,6 @@
                     src="images/drawer_open.png" title="Open" /><br />Open
                 </a></li>
                 <li wicket:id="showRemergeDocumentDialog" />
-                <li><a wicket:id="showExportDialog"> <img
-                    src="images/document_export.png" title="Export"  /><br />Export
-                </a></li>
               </ul>
             </fieldset>
           </li>
@@ -163,6 +159,7 @@
               <div class="panel-body flex-v-container" style="padding: 0px;">
                 <nav class="action-bar">
                   <span wicket:id="documentNavigator" />
+                  <span wicket:id="exportDialog" />
                   <span wicket:id="pageNavigator"/>
                   <span wicket:id="toggleScriptDirection"/>
                   <span wicket:id="guidelinesDialog" />
